requirements-dev.txt
black
pre-commit
pytest-playwright
PyGithub
